Move AddCloudFoundrySecurity inside AddCloudFoundryActuator
