+++ src/Management/src/Endpoint/Actuators/All/EndpointServiceCollectionExtensions.cs
@@ -56,7 +56,6 @@ public static class EndpointServiceCollectionExtensions
         if (Platform.IsCloudFoundry)
         {
             services.AddCloudFoundryActuator();
-            services.AddCloudFoundrySecurity();
         }
 
         services.AddHypermediaActuator();
