+++ src/Management/src/Endpoint/Actuators/CloudFoundry/EndpointServiceCollectionExtensions.cs
@@ -3,6 +3,9 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+using Microsoft.Extensions.Options;
+using Steeltoe.Common.Http.HttpClientPooling;
 using Steeltoe.Management.Endpoint.Actuators.Hypermedia;
 
 namespace Steeltoe.Management.Endpoint.Actuators.CloudFoundry;
@@ -26,6 +29,35 @@ public static class EndpointServiceCollectionExtensions
             .AddCoreActuatorServicesAsSingleton<CloudFoundryEndpointOptions, ConfigureCloudFoundryEndpointOptions, CloudFoundryEndpointMiddleware,
                 ICloudFoundryEndpointHandler, CloudFoundryEndpointHandler, string, Links>();
 
+        AddCloudFoundrySecurity(services);
+
         return services;
+    }
+
+    private static void AddCloudFoundrySecurity(IServiceCollection services)
+    {
+        services.AddSingleton<PermissionsProvider>();
+        ConfigureHttpClient(services);
+    }
+
+    private static void ConfigureHttpClient(IServiceCollection services)
+    {
+        services.TryAddSingleton<HttpClientHandlerFactory>();
+        services.TryAddSingleton<ValidateCertificatesHttpClientHandlerConfigurer<CloudFoundryEndpointOptions>>();
+
+        IHttpClientBuilder httpClientBuilder = services.AddHttpClient(PermissionsProvider.HttpClientName);
+
+        httpClientBuilder.ConfigurePrimaryHttpMessageHandler(serviceProvider =>
+        {
+            var handlerFactory = serviceProvider.GetRequiredService<HttpClientHandlerFactory>();
+            HttpClientHandler handler = handlerFactory.Create();
+
+            var validateCertificatesHandler =
+                serviceProvider.GetRequiredService<ValidateCertificatesHttpClientHandlerConfigurer<CloudFoundryEndpointOptions>>();
+
+            validateCertificatesHandler.Configure(Options.DefaultName, handler);
+
+            return handler;
+        });
     }
 }
