+++ src/Management/src/Endpoint/HostBuilderWrapperExtensions.cs
@@ -111,12 +111,7 @@ internal static class HostBuilderWrapperExtensions
 
     public static void AddCloudFoundryActuator(this HostBuilderWrapper wrapper)
     {
-        wrapper.ConfigureServices(services =>
-        {
-            services.AddCloudFoundryActuator();
-            services.AddCloudFoundrySecurity();
-        });
-
+        wrapper.ConfigureServices(services => services.AddCloudFoundryActuator());
         RegisterActuatorEndpoints(wrapper, null);
     }
